// client/src/_recoil/keys.ts
export enum EDialog {
    HANDLE_CREATE_DECK_DIALOG = 'HANDLE_CREATE_DECK_DIALOG',
    HANDLE_CREATE_CARD_DIALOG = 'HANDLE_CREATE_CARD_DIALOG',
    HANDLE_CREATE_CARD_CATEGORY_DIALOG = 'HANDLE_CREATE_CARD_CATEGORY_DIALOG',
    HANDLE_STUDY_MODE_DIALOG = 'HANDLE_STUDY_MODE_DIALOG',
    HANDLE_SIDE_NAV_BAR = 'HANDLE_SIDE_NAV_BAR',
    HANDLE_SIGN_UP_DIALOG = 'HANDLE_SIGN_UP_DIALOG',
    HANDLE_LOGIN_DIALOG = 'HANDLE_LOGIN_DIALOG',
}

export enum EAccount {
    HANDLE_LOGIN = 'HANDLE_LOGIN',
}

export enum EDeck {
    SET_ALL_DECK = 'SET_ALL_DECK',
    SET_DECK_ID = 'SET_DECK_ID',
    SET_DECK_INFO = 'SET_DECK_INFO',
    SET_DECK_NAME = 'SET_DECK_NAME',
    SET_STUDY_CARDS = 'SET_STUDY_CARDS',
}

export enum ECard {
    SET_ALL_CARD_CATEGORIES = 'SET_ALL_CARD_CATEGORIES',
    SET_TARGET_CARD = 'SET_TARGET_CARD',
    SET_TARGET_CARDS_COLS = 'SET_TARGET_CARDS_COLS',
}